+++ Testing/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 - Updated dependency on System.Buffers to 4.5.1
+- Internal API changes
 
 ## [2.2.2] - 2020-03-04